requirements.txt
fastapi
uvicorn
python-multipart
nltk
transformers>=4.40.0
torch
sacremoses
sentencepiece
accelerate
bitsandbytes
pydantic
python-docx
reportlab
beautifulsoup4
lxml
docx2pdf
